Handle case where SubnetId is not specified or is invalid
I hadn't tested the case where SubnetId isn't in the config file.
Had to move the SubnetId specific configuration after the creation of
the VPC object so could get the private subnets so I can choose a default
subnet from the list.

Added a check and error if the VPC doesn't have any private subnets.

Resolves #61