Set allow_all_outbound, allow_all_ipv6_outbound correctly when updating existing security group.
ec2.SecurityGroup.from_security_group_id is used to create a SecurityGroup resource from an existing security group so that it can be updated with additional security group rules.
This is done for the following security group types:
* config['slurm']['storage']['ExtraMountSecurityGroups']
* config['slurm']['ExistingSlurmDbd']['SecurityGroup']
* config['slurm']['Federation']['SlurmCtlSecurityGroups']
* config['slurm']['Federation']['SlurmNodeSecurityGroups']
* config['slurm']['SubmitterSecurityGroupIds']

The allow_all_outbound property was being set to False and the allow_all_ipv6_outbound was defaulting to True without consideration of the actual egress rules.
If all outbound traffic is being allowed the addition of egress rules removes
the rule and can break resources that rely on the unrestricted egress.

Use boto3 to check the actual egress rules and set the properties correctly
if rule are found that allow all egress traffic.

Verified that egress rules are no longer added if all traffic is allowed.

Resolves #72