Add SlurmConfOverrides config
Allow overrides to be added at end of slurm.conf

Change the order of slurm_nodes.conf to sort nodes by weight and to put
spot nodes first.
This partially for better organization and partly to see if putting spot
nodes first will let the scheduler pick them since it appears to be
ignoring weight when scheduling powered down nodes.

Resolves #99