Fix use of hpc6a instances
The plugin was seeing that the instance type is x86 and can have 2 threads
and was setting CpuOptions to set the number of threads per core to 1.
But this instance family has only 1 instance size and it disables hyperthreading
and doesn't allow using CpuOptions.
It also doesn't have spot pricing as an option so need to check if spot
pricing exists when creating spot compute nodes.

Update instance_type_info.json to save DefaultThreadsPerCore instead of
CoreCount.
Sync EC2InstanceTypeInfo with the copy in hpc-cost-simulator.
Replaces ThreadsPerCore.
Save ValidThreadsPerCore.
As part of the sync, rename instance_type_info to instance_type_and_family_info
because in hpc-cost-simulator it contains both.
This removes the need to create instance_family_info.
Adds savings plan info in pricing which slows things down and requires
additional IAM permissions for the controller.

Resolves #122