Enable ParallelCluster access from outside the cluster
Create CfnOutputs and scripts to configure a submission host to access
a ParallelCluster without having to ssh to it.

Install the same version of munge as used by ParallelCluster.

Copy ParallelCluster's munge key to the submission host.
In the future will test changing ParallelCluster to use a key from an
SSM parameter.

Create a command to mount the Slurm shared file system on the submission host.

Create a command to run a script to configure the submission host.

Create a modulefile to configure the environment to use the ParallelCluster
from the submission host.

Resolves #136