Updates to allow submitters to use multiple ParallelCluster clusters at the same time
Update the slurm config to add the ClusterName to the path.

Enable multiple clusters to be used by submitters

Change the mount point to include the cluster name so it is unique between clusters.
Updates scripts and config paths to include the cluster name in the paths.
Add a symbolic link so the head and compute nodes have access to the same
path as the submitter.

Resolves #139