Config arguments not being passed with correct types
When trying to update an existing cluster the head not root volume size is
being flagged becuase it is str(200), not int(200).
It turns out that when passing the config to the CloudFormation CustomResource
for the cluster it isn't preserving the integer types and is passing their
values as strings.

Pass the configuration to the CustomResource as a json string and parse the json
string back into a Python structure so that int and string types are correctly
preserved.

Also combine the custom resources for the S3 config objects and the
ParallelCluster so that a change to the json config string will be
assured to also trigger an updates of the cluster.

Move the customer resource lambda to create_parallel_cluster_lambdas()
where it belongs.

Save the config as json and yaml in a fixed s3 key in the assets bucket.

Delete CreateParallelClusterConfig lambda.
Functionality is move the CreateParallelClusterConfig.

Updates weren't happening if the fleet was running.
Added checking for this condition and stop the fleet and retry the update
when necessary.

Resolves #158