Don't pass config as property to CreateParallelCluster lambda
Create a jinja2 template of the config file and save it to s3.
Pass the template variable values as parameters to the lambda.
Get the template in the lambda and render it with the template parameters.
Save the rendered config to s3 and create/update the cluster.

Remove PluginDir from the custom config settings because it breaks plugin loading for different OS versions.

Resolves #179