Relax PyYAML version requirement
Not sure why it was requiring a specific version, but doesn't seem to be necessary.

Update setup.sh

Don't deactivate virtual venv because it can be setting up newer python version.

Use return instead of exit so don't exit calling shell.

Use "which node" instead of "node -v" to detect if nodejs is in the path.
"node-v" prompts to install nodejs, which I don't want because I want
control over which version is installed.

Add to documentation on how to use newer versions of python.

Resolves #212