Add support for ParallelCluster versions 3.9.0 and 3.9.1
Add support for rhel9 and rocky9.
Had to update some of the ansible playbooks to mimic rhel8 changes.

Resolves #229

Set SubmitterInstanceTags based on RESEnvironmentName.

Remove SubmitterSecurityGroupIds parameter.
This option added rules to existing security groups and if they were used by multiple clusters then the number of security group rules would exceed the maximum allowed.
With the addition of adding security groups to the head and compute nodes the
customer should supply their own security groups that meet the slurm cluster requirements, attach them to their login nodes and configure them as additional security groups for the head and compute nodes.

Resolves #204

Update CallSlurmRestApiLambda from Python 3.8 to 3.9.

Resolves #230

Update CDK version to 2.111.0.
This is the latest version supported by nodejs 16.
Really need to move to nodejs 20, but it isn't supported on Amazon Linux 2 or
RHEL 7 family.
Would require either running in a Docker container or on a newer OS version.
I think that I'm going to change the prerequisites for the OS distribution
so that I can stay on the latest tools.
For example, I can't update to Python 3.12 until I do this.

Update DeconfigureRESUsersGroupsJson to pass if last statement fails.

Fix bug in create_slurm_accounts.py

Resolves #231