Do not auto-prune instance types if there are too many
I was previously only allowing 1 memory size/core count combination to keep
the number of compute resources down and also was combining multiple instance
types in one compute resource if possible.
This was to try to maximize the number of instance types that were configured.

This led to people not being able to configure the exact instance types they
wanted.
The preference is to notify the user and let them choose which instances types
to exclude or to reduce the number of included types.

So, I've reverted to my original strategy of 1 instance type per compute resource and 1 CR per queue.
The compute resources can be combined into any queues that the user wants using
custom slurm settings.

I had to exclude instance types in the default configuration in order to keep from exceeding the PC limits.

Resolves #220

Update ParallelCluster version in config files and docs.

Clean up security scan.