Only use default EDA excludes if also using default Includes
Default Exclude only works correctly with default includes because it excludes
instance types to keep the total instance types down.
If user specifies any includes, then the default EDA excludes may exclude
instance types that they are trying to include.
Only use the default EDA includes and excludes if no includes or excludes
are configured.

Clean up the defaults to be less opinionated.
The EDA configuration is opinionated and meant to be a starting point.
But the defaults are wide open and will fail because they select too many instance types.
So, the errors will give guidance on winnowing down the instance types.

Resolves #262

Restore memory based partitions.

Related to #235.

Create partitions that include the purchase option (sp|od) and the amount of instance memory.
This maintains backward compantibility for those using partitions to select
the purchase option and amount of total memory.

Resolves #261

Add UseOnDemand configuration option that defaults to true so that you could
create a cluster with only spot instances.
Currently you always had to configure on-demand instances and only spot was optional.
Make both purchase options optional, but require at least one.