Name compute resources with the amount of memory and number of cores
Instead of naming the queue and the CR the same, after the instance type,
name the CR with the instance type memory and core amounts to give cluster
users more information about the compute nodes.
Many users will not know how much memory and how many cores are available
based on the instance type name.

Resolves #264