Allow additional CustomActions scripts for head and compute nodes.
For the head node add 2 scripts that can be edited on the head node that will be run when the head node is updated.

Do the same for the OnNodeStart and OnNodeConfigured custom actions for compute nodes.

Add additional parameter to provide additional compute node CustomActions in the ParallelCluster config.

Resolves #320